refactor(lsp): separate builder and parser
