+++ packages/language-server/src/model/index.ts
@@ -1,4 +1,5 @@
-export * from './fqn-computation';
-export * from './fqn-index';
-export * from './model-builder';
-export * from './model-locator';
+export * from './fqn-computation'
+export * from './fqn-index'
+export * from './model-builder'
+export * from './model-locator'
+export * from './model-parser'
