+++ packages/vscode/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../tsconfig/base.json",
   "compilerOptions": {
     "lib": ["ESNext", "WebWorker"],
-    "skipLibCheck": true,
     "paths": {
       "src/*": ["./src/*"]
     }
